Island: Extract AuthenticationFacade construction
Extract AuthenticationFacade construction from setup_authentication
